+++ src/pages/messages/create-chat/create-chat.styles.scss
@@ -17,6 +17,7 @@ html {
     background-color: #fefefe;
 
     margin: 0 auto;
+    margin-top: 30vh;
 
     > * {
         width:80%;
@@ -55,13 +56,12 @@ html {
 
 .create-modal {
     display: none;
-    position: fixed; /* Stay in place */
-    z-index: 1; /* Sit on top */
+    position: fixed;
+    z-index: 1;
     left: 0;
     top: 0;
-    width: 100%; /* Full width */
-    height: 100%; /* Full height */
-    overflow: auto; /* Enable scroll if needed */
-    background-color: rgb(0,0,0); /* Fallback color */
-    background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
+    width: 100%;
+    height: 100%;
+    overflow: auto;
+    background-color: rgba(0,0,0,0.4);
 }
