+++ src/pages/messages/create-chat/create-chat.template.pug
@@ -1,5 +1,5 @@
 div(class='create-modal' id='create-modal')
     div(class='create-modal-container')
         h3(class='create-chat-label') Введите название чата
-        input(class='create-chat-input' type='text')
+        input(class='create-chat-input' id='create-chat-input' type='text')
         button(class='create-chat-button' id='create-chat-button') Создать
